+++ app/src/main/java/com/paulsoja/depositcalculator/Activity/MainActivity.java
@@ -1,7 +1,10 @@
-package com.paulsoja.depositcalculator;
+package com.paulsoja.depositcalculator.Activity;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+
+import com.paulsoja.depositcalculator.Fragments.InputDepositDataFragment;
+import com.paulsoja.depositcalculator.R;
 
 public class MainActivity extends AppCompatActivity{
