+++ app/src/main/res/layout-large-land/activity_main.xml
@@ -5,11 +5,11 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="horizontal"
-    tools:context="com.paulsoja.depositcalculator.MainActivity">
+    tools:context="com.paulsoja.depositcalculator.Activity.MainActivity">
 
     <fragment
         android:id="@+id/Fragment_inputDepositData"
-        android:name="com.paulsoja.depositcalculator.InputDepositDataFragment"
+        android:name="com.paulsoja.depositcalculator.Fragments.InputDepositDataFragment"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="1"
@@ -17,7 +17,7 @@
 
     <fragment
         android:id="@+id/Fragment_viewResultDeposit"
-        android:name="com.paulsoja.depositcalculator.ViewResultDepositFragment"
+        android:name="com.paulsoja.depositcalculator.Fragments.ViewResultDepositFragment"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_weight="1"
